wechatsell/src/main/java/com/shy/wechatsell/service/impl/CategoryServiceImpl.java
package com.shy.wechatsell.service.impl;

import com.shy.wechatsell.dataobject.ProductCategory;
import com.shy.wechatsell.repository.ProductCategoryRepository;
import com.shy.wechatsell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author Haiyu
 * @date 2018/10/15 13:25
 */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository repository;
    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public ProductCategory findOne(Integer id) {
        return repository.findById(id).get();
    }

    @Override
    public List<ProductCategory> findByCategoryTypes(List<Integer> categoryList) {
        return repository.findByCategoryTypeIn(categoryList);
    }

    @Override
    public ProductCategory save(ProductCategory category) {
        return repository.save(category);
    }
}
